Move TagAliasGroup into the impl package
